refactor: modularize socket io
